+++ console/src/app/pages/orgs/orgs-routing.module.ts
@@ -6,11 +6,6 @@ import { OrgCreateComponent } from './org-create/org-create.component';
 import { OrgDetailComponent } from './org-detail/org-detail.component';
 import { OrgGridComponent } from './org-grid/org-grid.component';
 
-export enum PolicyComponentAction {
-    CREATE = 'create',
-    MODIFY = 'modify',
-}
-
 const routes: Routes = [
     {
         path: 'create',
@@ -30,60 +25,29 @@ const routes: Routes = [
         },
     },
     {
-        path: 'policy/age',
-        data: {
-            action: PolicyComponentAction.MODIFY,
-        },
-        loadChildren: () => import('./password-age-policy/password-age-policy.module')
-            .then(m => m.PasswordAgePolicyModule),
-    },
-    {
-        path: 'policy/lockout/create',
-        data: {
-            action: PolicyComponentAction.CREATE,
-        },
-        loadChildren: () => import('./password-lockout-policy/password-lockout-policy.module')
-            .then(m => m.PasswordLockoutPolicyModule),
-    },
-    {
-        path: 'policy/lockout',
-        data: {
-            action: PolicyComponentAction.MODIFY,
-        },
-        loadChildren: () => import('./password-lockout-policy/password-lockout-policy.module')
-            .then(m => m.PasswordLockoutPolicyModule),
-    },
-    {
-        path: 'policy/complexity/create',
-        data: {
-            action: PolicyComponentAction.CREATE,
-        },
-        loadChildren: () => import('./password-complexity-policy/password-complexity-policy.module')
-            .then(m => m.PasswordComplexityPolicyModule),
-    },
-    {
-        path: 'policy/complexity',
-        data: {
-            action: PolicyComponentAction.MODIFY,
-        },
-        loadChildren: () => import('./password-complexity-policy/password-complexity-policy.module')
-            .then(m => m.PasswordComplexityPolicyModule),
-    },
-    {
-        path: 'policy/iam_policy/create',
-        data: {
-            action: PolicyComponentAction.CREATE,
-        },
-        loadChildren: () => import('./password-iam-policy/password-iam-policy.module')
-            .then(m => m.PasswordIamPolicyModule),
-    },
-    {
-        path: 'policy/iam_policy',
-        data: {
-            action: PolicyComponentAction.MODIFY,
-        },
-        loadChildren: () => import('./password-iam-policy/password-iam-policy.module')
-            .then(m => m.PasswordIamPolicyModule),
+        path: 'policy',
+        children: [
+            {
+                path: 'age',
+                loadChildren: () => import('./policies/password-age-policy/password-age-policy.module')
+                    .then(m => m.PasswordAgePolicyModule),
+            },
+            {
+                path: 'lockout',
+                loadChildren: () => import('./policies/password-lockout-policy/password-lockout-policy.module')
+                    .then(m => m.PasswordLockoutPolicyModule),
+            },
+            {
+                path: 'complexity',
+                loadChildren: () => import('./policies/password-complexity-policy/password-complexity-policy.module')
+                    .then(m => m.PasswordComplexityPolicyModule),
+            },
+            {
+                path: 'iam_policy',
+                loadChildren: () => import('./policies/password-iam-policy/password-iam-policy.module')
+                    .then(m => m.PasswordIamPolicyModule),
+            },
+        ],
     },
     {
         path: 'members',
