+++ console/src/app/pages/orgs/policies/password-age-policy/password-age-policy-routing.module.ts
@@ -1,13 +1,25 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
+import { PolicyComponentAction } from '../../orgs-routing.module';
 import { PasswordAgePolicyComponent } from './password-age-policy.component';
 
 const routes: Routes = [
     {
         path: '',
         component: PasswordAgePolicyComponent,
-        data: { animation: 'DetailPage' },
+        data: {
+            animation: 'DetailPage',
+            action: PolicyComponentAction.MODIFY,
+        },
+    },
+    {
+        path: 'create',
+        component: PasswordAgePolicyComponent,
+        data: {
+            animation: 'DetailPage',
+            action: PolicyComponentAction.CREATE,
+        },
     },
 ];
