+++ console/src/app/pages/orgs/policies/password-age-policy/password-age-policy.component.scss
