+++ console/src/app/pages/orgs/policies/password-complexity-policy/password-complexity-policy-routing.module.ts
@@ -1,13 +1,25 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
+import { PolicyComponentAction } from '../policy-component-action.enum';
 import { PasswordComplexityPolicyComponent } from './password-complexity-policy.component';
 
 const routes: Routes = [
     {
         path: '',
         component: PasswordComplexityPolicyComponent,
-        data: { animation: 'DetailPage' },
+        data: {
+            animation: 'DetailPage',
+            action: PolicyComponentAction.MODIFY,
+        },
+    },
+    {
+        path: 'create',
+        component: PasswordComplexityPolicyComponent,
+        data: {
+            animation: 'DetailPage',
+            action: PolicyComponentAction.CREATE,
+        },
     },
 ];
