+++ console/src/app/pages/orgs/policies/password-lockout-policy/password-lockout-policy-routing.module.ts
@@ -1,13 +1,25 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
+import { PolicyComponentAction } from '../policy-component-action.enum';
 import { PasswordLockoutPolicyComponent } from './password-lockout-policy.component';
 
 const routes: Routes = [
     {
         path: '',
         component: PasswordLockoutPolicyComponent,
-        data: { animation: 'DetailPage' },
+        data: {
+            animation: 'DetailPage',
+            action: PolicyComponentAction.MODIFY,
+        },
+    },
+    {
+        path: 'create',
+        component: PasswordLockoutPolicyComponent,
+        data: {
+            animation: 'DetailPage',
+            action: PolicyComponentAction.CREATE,
+        },
     },
 ];
