+++ console/src/app/pages/orgs/policies/password-lockout-policy/password-lockout-policy.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnDestroy } from '@angular/core';
+import { FormGroup } from '@angular/forms';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
@@ -11,15 +12,14 @@ import {
 import { ManagementService } from 'src/app/services/mgmt.service';
 import { ToastService } from 'src/app/services/toast.service';
 
-import { PolicyComponentAction } from '../orgs-routing.module';
-
+import { PolicyComponentAction } from '../policy-component-action.enum';
 
 @Component({
-    selector: 'app-password-policy',
-    templateUrl: './password-complexity-policy.component.html',
-    styleUrls: ['./password-complexity-policy.component.scss'],
+    selector: 'app-password-lockout-policy',
+    templateUrl: './password-lockout-policy.component.html',
+    styleUrls: ['./password-lockout-policy.component.scss'],
 })
-export class PasswordComplexityPolicyComponent implements OnDestroy {
+export class PasswordLockoutPolicyComponent implements OnDestroy {
     public title: string = '';
     public desc: string = '';
 
@@ -27,8 +27,8 @@ export class PasswordComplexityPolicyComponent implements OnDestroy {
 
     public PolicyComponentAction: any = PolicyComponentAction;
 
-    public complexityData!: PasswordComplexityPolicy.AsObject;
-
+    public lockoutForm!: FormGroup;
+    public lockoutData!: PasswordLockoutPolicy.AsObject;
     private sub: Subscription = new Subscription();
 
     constructor(
@@ -41,13 +41,13 @@ export class PasswordComplexityPolicyComponent implements OnDestroy {
             this.componentAction = data.action;
             return this.route.params;
         })).subscribe(params => {
-            this.title = 'ORG.POLICY.PWD_COMPLEXITY.TITLECREATE';
-            this.desc = 'ORG.POLICY.PWD_COMPLEXITY.DESCRIPTIONCREATE';
+            this.title = 'ORG.POLICY.PWD_LOCKOUT.TITLECREATE';
+            this.desc = 'ORG.POLICY.PWD_LOCKOUT.DESCRIPTIONCREATE';
 
             if (this.componentAction === PolicyComponentAction.MODIFY) {
                 this.getData(params).then(data => {
                     if (data) {
-                        this.complexityData = data.toObject() as PasswordComplexityPolicy.AsObject;
+                        this.lockoutData = data.toObject() as PasswordLockoutPolicy.AsObject;
                     }
                 });
             }
@@ -60,56 +60,49 @@ export class PasswordComplexityPolicyComponent implements OnDestroy {
 
     private async getData(params: any):
         Promise<PasswordLockoutPolicy | PasswordAgePolicy | PasswordComplexityPolicy | OrgIamPolicy | undefined> {
-        this.title = 'ORG.POLICY.PWD_COMPLEXITY.TITLE';
-        this.desc = 'ORG.POLICY.PWD_COMPLEXITY.DESCRIPTION';
-        return this.mgmtService.GetPasswordComplexityPolicy();
+
+        this.title = 'ORG.POLICY.PWD_LOCKOUT.TITLE';
+        this.desc = 'ORG.POLICY.PWD_LOCKOUT.DESCRIPTION';
+        return this.mgmtService.GetPasswordLockoutPolicy();
     }
 
     public deletePolicy(): void {
-        this.mgmtService.DeletePasswordComplexityPolicy(this.complexityData.id).then(() => {
+        this.mgmtService.DeletePasswordLockoutPolicy(this.lockoutData.id).then(() => {
             this.toast.showInfo('Successfully deleted');
         }).catch(error => {
             this.toast.showError(error);
         });
     }
 
-    public incrementLength(): void {
-        if (this.complexityData?.minLength !== undefined && this.complexityData?.minLength <= 72) {
-            this.complexityData.minLength++;
+    public incrementMaxAttempts(): void {
+        if (this.lockoutData?.maxAttempts !== undefined) {
+            this.lockoutData.maxAttempts++;
         }
     }
 
-    public decrementLength(): void {
-        if (this.complexityData?.minLength && this.complexityData?.minLength > 1) {
-            this.complexityData.minLength--;
+    public decrementMaxAttempts(): void {
+        if (this.lockoutData?.maxAttempts && this.lockoutData?.maxAttempts > 0) {
+            this.lockoutData.maxAttempts--;
         }
     }
 
     public savePolicy(): void {
         if (this.componentAction === PolicyComponentAction.CREATE) {
-
-            this.mgmtService.CreatePasswordComplexityPolicy(
-                this.complexityData.description,
-                this.complexityData.hasLowercase,
-                this.complexityData.hasUppercase,
-                this.complexityData.hasNumber,
-                this.complexityData.hasSymbol,
-                this.complexityData.minLength,
+            this.mgmtService.CreatePasswordLockoutPolicy(
+                this.lockoutData.description,
+                this.lockoutData.maxAttempts,
+                this.lockoutData.showLockOutFailures,
             ).then(() => {
                 this.router.navigate(['org']);
             }).catch(error => {
                 this.toast.showError(error);
             });
-
         } else if (this.componentAction === PolicyComponentAction.MODIFY) {
 
-            this.mgmtService.UpdatePasswordComplexityPolicy(
-                this.complexityData.description,
-                this.complexityData.hasLowercase,
-                this.complexityData.hasUppercase,
-                this.complexityData.hasNumber,
-                this.complexityData.hasSymbol,
-                this.complexityData.minLength,
+            this.mgmtService.UpdatePasswordLockoutPolicy(
+                this.lockoutData.description,
+                this.lockoutData.maxAttempts,
+                this.lockoutData.showLockOutFailures,
             ).then(() => {
                 this.router.navigate(['org']);
             }).catch(error => {
