+++ console/src/app/pages/orgs/policies/policy-component-action.enum.ts
@@ -1,0 +1,5 @@
+
+export enum PolicyComponentAction {
+    CREATE = 'create',
+    MODIFY = 'modify',
+}
